Analysis stage (#844)
This PR reworks the chainsync data to dashboard pipeline to move
computation away from the dashboard and into a separate process.

Major Changes:
- New `run_data_analysis.py` process to run analysis after
`acquire_data` that focuses on computation.
- This stage currently computes spot price, fixed rate, base buffer,
wallet positions, pnl, and the ticker.
- Adding new tables (and interface) to support output of data analysis.
- Simplifying dashboard frontend to query output of data analysis.

Minor Changes:
- Adding drop argument to sqlalchemy `initialize_session` to drop
existing tables if set to True for debugging.
- Adding in create table retries due to race condition between acquire
data and data analysis both trying to create tables.
- Removing obsolete agent positions data class.
- Renames WalletInfo schema to be WalletInfoFromChain.
- Removing chainsync scripts for calculating leaderboards from previous
trading competition.
- Renaming table names to be snake case.
- Adding in a rerun of bots + acquire data + data analysis in system
test to test data pipeline restarts.
- Splitting out system tests to different files.

Breaking Change:
- postgres db interface getter functions no longer return blockNumber as
an index. This is to keep the getter functions consistent, as not every
table stores the block number.

Bug Fixes:
- `smart_contract_preview_transaction` now takes in a block number
argument to query the mock trade at that exact block. This helps solve a
race condition where the preview could be mocking a position that's no
longer there.

TODO:
- More plots in dashboard.