consistent naming of URIs (#890)
Web3 uses the term "URI" instead of "URL", so this PR renames our
variables to be consistent. The two terms are actually [not the
same](https://en.wikipedia.org/wiki/Uniform_Resource_Identifier). A URL
is a type of URI that specifies the asset (also known as a URN) and
resource (for example, `http://`). Since we may not need to specify a
resource, we use the more general "URI" term.

99% of the time we will be providing a URL and the distinction is all
but pedantic. It only starts to matter if we say e.g. `my_url: URI =
URI("address")` -- this is immediately contradictory, which occurs when
we are creating variables that have the web3py `URI` type.