add hyperdrive types (#888)
This PR adds the `hyperdrive_types` package to `lib`. This package is
comprised of a shell script that recursively parses an `abi` folder for
ABI JSON files, and calls `pypechain` to generate the corresponding
python files. Due to problems described in the discussion below, we
opted to also commit the programmatically generated python files to this
repo.

Original description of the problem is in [this comment
below](https://github.com/delvtech/elf-simulations/pull/888#issuecomment-1731833315).