add a hyperdrive api (#870)
This PR is to create a higher-level interface for the hyperdrive
contract. Presently, the primary consumer of this API will be agent0.

The api is primarily following specs from the front-end team's
`hyperdrive-js` sdk. It will ultimately wrap
- pypechain: types, errors, structs, etc
- pyperdrive: static functions such as get_max_long, get_spot_price, etc
- ethpy.hyperdrive: more convenient interface for trades (open_long,
etc), getting the pool state & config, etc

This PR is a first step, and does not include all of the desired
features.

The API currently provides the following endpoints:
- hyperdrive.web3
- hyperdrive.hyperdrive_contract
- hyperdrive.base_token_contract
- hyperdrive.get_get_eth_base_balances(address)
- hyperdrive.last_state_block_number
- hyperdrive.pool_config
- hyperdrive.pool_info
- hyperdrive.latest_checkpoint
- hyperdrive.current_block
- hyperdrive.current_block_number
- hyperdrive.current_block_time
- hyperdrive.spot_price
- hyperdrive.get_max_long(FixedPoint(1000))
- hyperdrive.get_max_short(FixedPoint(1000))