replace elfpy market with ethpy HyperdriveInterface (#909)
This PR modifies the agent0 policies & accounts to use the new
HyperdriveInterface API instead of the elfpy HyperdriveMarket.

The new interface lacks some features (e.g. all of the yieldspace math
and variable rates) that were required for the smart long and short
policies, so we are deleting those for now. I made several new issues to
detail follow-up work that is still needed. They are #910 #911 #913

This PR also gets us very close to completely detaching elfpy from the
rest of the monorepo, though the rest of the detachment should also
happen in a follow-up PR.