Address -> username and username -> user rework (#975)
Reworking the way the dashboard is getting the address to username and
the username to user mappings.

Address -> username maps many addresses to a single username. For
example, I have 3 wallets mapped to a single bot username.
username -> user maps many usernames to a single user. For example, my
bot account and click account gets mapped to a single user for the
leaderboard.

These mappings are written to the database and can be live updated with
the `update_username.py` script in `lib/chainsync/bin`. This script also
gets ran during docker initialization (corresponding infra change
needed).