Funding bots now async with retries (#983)
This PR allows for funding eth, base, and base approvals with bots in
async. This allows setup of bots on a chain mined via time quickly, with
a large number of bots.

Part of this was a refactor to smart_contract_write calls to allow for a
manual nonce parameter. This is needed for funding, as we launch
multiple transactions at the same time. The previous method of getting
the transaction count from web3 resulted in nonce errors.

The above also fixes an issue when an agent launches multiple trades on
the same block.

Also fixes logging to the `.logging` directory