invariant tests (#1138)
This is a WIP and establishes the framework for the tests outlined
below. More work is required to finish up the tests and set up a system
for running them on cloud chain.

Adds an invariant test script to be run alongside fuzz bots. It checks
the following:

- total shares == share reserves + shorts outstanding/c +
_governanceFeesAccrued + _withdrawalPool.proceeds + epsilon
- the system should always be solvent (solvency = share_reserves -
global_exposure - minimum_share_reserves > 0)
- longs and shorts should always be closed at maturity for the correct
amounts
-  share reserves * c - long exposure >= minimum share reserves
- balanceOf(hyperdrive) == share reserves + shorts outstanding/c +
_governanceFeesAccrued + _withdrawalPool.proceeds + epsilon

This check is additionally added if the checkpoint bot script is
executed with the `fuzz` flag set.

-  creating a checkpoint should never fail