split hyperdrive interface into read & readwrite versions (#1220)
The Hyperdrive Interface object has read operations (like getting block
information, or pool info) and write operations (like opening a long or
short). We want to be able to restrict the type to read-only or
read-write versions of the interface to remove any ambiguity about who
should and should not be executing trades. Specifically, the bot
policies all share an `action` function that is expected to return a
list of Trades. The action function is not supposed to execute trades
itself, since that is executed elsewhere to account for database
updates, error handling, etc. Now all policies specifically receive the
ReadInterface to remove any ambiguity about the allowed action set.

I decided to use an inheritance model instead of class attributes, which
is how I originally proposed it in the
[issue](https://github.com/delvtech/agent0/issues/1128). I decided on
this because it removes a step when importing & declaring types
(`HyperdriveInterface.ReadInterface` versus only
`HyperdriveReadInterface`).