Use pypechain's new deploy method (#1218)
Use this instead of `deploy_contract`, which is now no longer used. We
can delete it or leave it since its still a utility that ethpy could
provide.

I've also totally removed the dependency of `abi_dir` in eth_config
throughout the repo... I think. I'm not taking the final step of
removing that either in this PR in case there is some use case I don't
know about.