refactor interface, fixtures, and agent state objects (#1306)
The goal of this PR was to remove any `ethpy` dependencies on `agent0`,
the primary example of which was in `HyperdriveReadInterface`. Fixing
this illuminated or caused a handful of other dependency and
organizational issues that I also resolved.

All changes are mechanical (moves, renames, adjusting imports and init
files), so the primary review I'm looking for is agreement with the new
organization and import structure.


# interface
* the trade static methods (e.g. `open_long_trade`) have been moved out
of interface and into `agent0/hyperdrive/hyperdrive_actions.py`; this
removes ethpy dependencies on agent0.

# test fixtures
* there were agent0 fixture objects in ethpy, those now live in agent0
* ethpy fixtures are split into a file for chain objects and a file for
pool objects
* agent0 now houses the interface fixtures and interactive chain
fixtures
* convtest fixture list is alphabetized

# state
* agent0 all class objects for the agent that used to live in
`agent0/hyperdrive/state/` now live in `agent0/hyperdrive/agent/`. Same
is true for base objects. This resolved a lot of dependencies issues,
where the agent and state were tightly interdependent.

# imports
* I moved some imports under the `TYPE_CHECKING` flag, although I did
not conduct an exhaustive review for this
* The following commonly-used objects can be directly imported `from
agent0.hyperdrive` instead of from submodules:
```
HyperdriveActionType
HyperdriveAgent
HyperdriveMarketAction
HyperdriveWallet
HyperdriveWalletDeltas
Long
Short
TradeResult
TradeStatus
HyperdriveReadInterface
HyperdriveReadWriteInterface
```

# renames
* `agent0/hyperdrive/agents` -> `agent0/hyperdrive/agent`
* `HyperdrivePolicy` -> `HyperdriveBasePolicy`
* `Zoo` -> `PolicyZoo`