refactor exec folder into separate test and IHyperdrive dependencies (#1362)
This PR is strictly a refactor; no new logic was implemented (see
caveats below).

A majority of the functions in the `agent0.core.exec` folder were only
used by tests. A smaller subset of them were used by `IHyperdrive`. I
split these up, and put the IHyperdrive ones in the `interactive`
folder. I consolidated the rest into `agent0.core.hyperdrive.utilities`,
which now also contains the predict files from `agent0.core.utilities`.

Now that we've severed the connection to interactive hyperdrive, the
functions located in `agent0.core.hyperdrive.utilities.run_bots` can be
paired down in a follow-up PR. Nearly all of the functions there are not
used anywhere else in the repo; strictly in tests. We should set up
example flows that would use them, and if we can't come up with any then
we should delete them and the corresponding tests. This will be done in
a follow-up PR, though.

caveats (functional changes):
- I made some functions 'private' (i.e. with an underscore in the name)
if they were only used by other functions inside the same file.
- I made `set_max_approval` non-async. It was not being used as an async
function in the first place because it was only used by [`IHyperdrive`
and always with the list length of
1](https://github.com/delvtech/agent0/blob/98288dadf6815921bdeb701fe7318ee4ba124ef4/src/agent0/core/hyperdrive/interactive/i_hyperdrive.py#L209).
This means as used there would always be only one thread. So I made it
non-async to remove unnecessary and unused code, with the idea that we
can make it async later if we have a use-case for it.
- the `get_agent_trades` flow was in a similar situation -- we had a
function for calling it with a list of agents but then hard-coded a
single `[agent]` input whenever it was used. I left parts async that
made sense to be (matching contract calls to trades), but the outer loop
would have to be made explicitly async in the application space (which
hasn't existed yet).