Remove unused config (#1462)
This PR cleans up various unnecessary configs, enabled by making the old
workflow obsolete.

- Removes obsolete environment config, agent config, and accounts
config.
- Removes the `EthConfig` object in favor of passing in an `rpc_uri` and
`hyperdrive_address` for the interface.
- A future refactor of deploying a registry can set a default of
`hyperdrive_address` to a specific pool.
- Removes several obsolete scripts specific for infra.
- We'll need to reimplement a checkpoint bot and fuzz bots for the infra
workflow, but we likely can use the existing checkpoint bot script for
testnet. We just need to deploy a registry in infra for this.
- Consolidating logging setup in interactive objects, and exposing
logging parameters in config.
- Exposing `preview_before_trade` up the call chain to avoid setting it
in `EthConfig`.
- Wrapping `redeem_withdraw_shares` preview call in
`preview_before_trade`.