Allow for lazy pnl calculation (#1561)
This PR allows for only calculating pnl when calling `get_positions` for
speed. To utilize this, set `calc_pnl = False` in the `Chain.Config`. If
this configuration is set to False, `get_positions` accepts a `calc_pnl`
argument, which fills the pnl values in the resulting dataframe.

Includes a bugfix to `get_positions` when getting positions from
multiple pools.